+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] / 2023-01-31
+### Features
+- [x] Add `Directory.Build.props` one file change all the versions with `Condition`
+- [x] Add Example Command ToolTip with Version
+- [x] Initialize only download with greater Version and return true if download
+- [x] Update `CommandUpdate` with `ShowBalloon`
+### Added
+- Add `InfoCenterUtil` with `ShowBalloon`
+
 ## [0.2.0] / 2023-01-31
 ### Features
 - [x] Initialize Auto Download Last Version - `Console log`
@@ -34,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add Example
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.3.0]: ../../compare/0.2.0...0.3.0
 [0.2.0]: ../../compare/0.1.0...0.2.0
 [0.1.0]: ../../compare/0.1.0
 [1.0.0]: ../../compare/1.0.0